Automatic Sync from SCM-svn to GitHub syncing the following commits: RosaBulo: Rosa Bulo (REB) SCMSUITE-7299 SO107
Made a change to from_smiles. If there is overlap between atoms in the created coordinates, then there will be a change of settings,
and new random coordinates will be created, which will generally not overlap.
I chose not to make this the default for from_smiles, because I would prefer the default generation to be none-random.